Use minor-mode macro initialization and buffer-local variables
When juggling helix-mode between separate buffers, there was a bug
where `helix--current-state' would persist when switching buffers
even if the value didn't match the actual minor mode active in that
buffer. With buffer-local variables, we no longer have that problem
because the value of `helix--current-state' is only persisted within
a single buffer.