Improved word navigation (#7)
* Modify forward-word and backward-word to mirror Helix

Try matching Helix behavior by navigating via re-search-forward
and re-search-backward using a regular expression instead of simply
forward/backward-word. The tough part is that re-search-backward
has different cursor behavior due to all Emacs regular expressions
searching in a forward direction, regardless of if the search is
specified as backward.

To compensate, there are some extra calls to `skip-syntax-backward`,
matching the pattern in the library function `forward-whitespace`.